integrate protractor into package scripts
