+++ packages/composer-playground/package.json
@@ -34,7 +34,8 @@
     "docs": "npm run typedoc -- --options typedoc.json --exclude '**/*.spec.ts' ./src/",
     "e2e:live": "npm run e2e -- --elementExplorer",
     "e2e": "npm run protractor",
-    "e2e:protractor": "npm run server:dev & npm run protractor:delay",
+    "e2e:nobuild": "node ./e2e/scripts/test.js",
+    "e2e:main": "npm run build:prod && node ./e2e/scripts/test.js",
     "github-deploy:dev": "webpack --config config/webpack.github-deploy.js --progress --profile --env.githubDev",
     "github-deploy:prod": "webpack --config config/webpack.github-deploy.js --progress --profile --env.githubProd",
     "github-deploy": "npm run github-deploy:dev",
@@ -50,7 +51,6 @@
     "pree2e": "npm run webdriver:update -- --standalone",
     "preversion": "npm test",
     "protractor": "protractor --protractor.conf.js",
-    "protractor:delay": "sleep 60 && npm run protractor --protractor.conf.js",
     "rimraf": "rimraf",
     "server:dev:hmr": "npm run server:dev -- --inline --hot",
     "server:dev": "webpack-dev-server --config config/webpack.dev.js --progress --profile --watch --content-base src/",
